\name{jsonRPCviaMQ-package}
\alias{jsonRPCviaMQ-package}
\alias{jsonRPCviaMQ}
\docType{package}
\title{Execute JSON-RPC messages retrieved from message queue}
\description{A very simple json-rpc server (loop) that gets requests from a message queue (activeMQ or rabbitMQ) and will reply to a queue if necessary.}
\details{
\tabular{ll}{Package: \tab jsonRPCviaMQ\cr
Type: \tab Package\cr
Title: \tab Execute JSON-RPC messages retrieved from message queue\cr
Version: \tab 0.5\cr
Date: \tab 2013-04-17\cr
Author: \tab Jesse Ross <jross@cornell.edu>\cr
Depends: \tab R (>= 2.13.0), rJava (>= 0.6-0), messageQueue (>= 0.4), RJSONIO, futile.logger, testthat\cr
Suggests: \tab testthat\cr
Imports: \tab rJava (>= 0.6-0), RJSONIO, messageQueue (>= 0.4), futile.logger\cr
SystemRequirements: \tab Java (>=7.0)\cr
Maintainer: \tab Matt MacGillivray <matthew.macgillivray@cornell.edu>\cr
License: \tab LGPL (>= 2.0)\cr
Repository: \tab r-forge\cr
URL: \tab http://jsonRPCviaMQ.r-forge.r-project.org/\cr}
}
\author{Jesse Ross <jross@cornell.edu>}

\keyword{ package }


