\name{jsonRPCviaMQ.echo}
\alias{jsonRPCviaMQ.echo}
\title{jsonRPCviaMQ echo}
\description{
Echo a request to the response queue.  This is a handy method for testing end-to-end.
}
\usage{jsonRPCviaMQ.echo(text)}
\arguments{
  \item{text}{
}
}



\author{Jesse Ross <jross@cornell.edu>}





