\name{jsonRPCviaMQ.executeRPC}
\alias{jsonRPCviaMQ.executeRPC}
\title{jsonRPCviaMQ executeRPC}
\description{
Execute a single RPC from an R object structured like a jsonrpc object
Returns a list with all parameters needed for a jsonrpc response

Error codes outlined here: http://www.jsonrpc.org/specification
}
\usage{jsonRPCviaMQ.executeRPC(call, exec_time)}
\arguments{
  \item{call}{structured variable representing the jsonrpc call}
  \item{exec_time}{
}
}



\author{Jesse Ross <jross@cornell.edu>}





