\name{jsonRPCviaMQ.runOnce}
\alias{jsonRPCviaMQ.runOnce}
\title{jsonRPCviaMQ runOnce}
\description{
Convenience method.

Connect to a queue, check for a single message and execute if a valid message was received.
Exit after checking for 1 message.
}
\usage{jsonRPCviaMQ.runOnce(queue, host, queueType)}
\arguments{
  \item{queue}{
}
  \item{host}{
}
  \item{queueType}{
}
}



\author{Jesse Ross <jross@cornell.edu>}





