\name{jsonRPCviaMQ.toJsonRpcError}
\alias{jsonRPCviaMQ.toJsonRpcError}
\title{jsonRPCviaMQ toJsonRpcError}
\description{
Helper method

create a JSON RPC error message
}
\usage{jsonRPCviaMQ.toJsonRpcError(id, error_code, error_message, version = "2.0")}
\arguments{
  \item{id}{
}
  \item{error_code}{
}
  \item{error_message}{
}
  \item{version}{
}
}



\author{Jesse Ross <jross@cornell.edu>}





