\name{jsonRPCviaMQ.toJsonRpcSuccess}
\alias{jsonRPCviaMQ.toJsonRpcSuccess}
\title{jsonRPCviaMQ toJsonRpcSuccess}
\description{
Helper method

create a JSON RPC success message
}
\usage{jsonRPCviaMQ.toJsonRpcSuccess(id, result = NULL, version = "2.0")}
\arguments{
  \item{id}{
}
  \item{result}{
}
  \item{version}{
}
}



\author{Jesse Ross <jross@cornell.edu>}





