\name{jsonRPCviaMQ.who}
\alias{jsonRPCviaMQ.who}
\title{jsonRPCviaMQ who}
\description{
Echo a request to the response queue.  This is a handy method for testing end-to-end.
}
\usage{jsonRPCviaMQ.who(sleepS = 10)}
\arguments{
  \item{sleepS}{how long to sleep after indicating that this machine is listening (allows other machines to respond)}
}



\author{Jesse Ross <jross@cornell.edu>}





